src/main/persistence/JsonReader.java
package persistence;

import model.LogHistory;
import org.json.JSONObject;

import java.io.IOException;

/**
 * Represents a reader that reads LogHistory from JSON data stored in file
 */
public class JsonReader {
    //private String source;

    // EFFECTS: constructs reader to read from source file
    public JsonReader(String source) {
        // TODO
        //this.source = source;
    }

    // EFFECTS: reads LogHistory from file and returns it;
    // throws IOException if an error occurs reading data from file
    public LogHistory read() throws IOException {
//        String jsonData = readFile(source);
//        JSONObject jsonObject = new JSONObject(jsonData);
//        return parseWorkRoom(jsonObject);
        return new LogHistory() {
            @Override
            public JSONObject toJson() {
                return null;
            }
        };
    }

    // TODO
    /*
    // EFFECTS: reads source file as string and returns it
    private String readFile(String source) throws IOException {
        StringBuilder contentBuilder = new StringBuilder();

        try (Stream<String> stream = Files.lines( Paths.get(source), StandardCharsets.UTF_8)) {
            stream.forEach(s -> contentBuilder.append(s));
        }

        return contentBuilder.toString();
    }

    // EFFECTS: parses workroom from JSON object and returns it
    private WorkRoom parseWorkRoom(JSONObject jsonObject) {
        String name = jsonObject.getString("name");
        WorkRoom wr = new WorkRoom(name);
        addThingies(wr, jsonObject);
        return wr;
    }

    // MODIFIES: wr
    // EFFECTS: parses thingies from JSON object and adds them to workroom
    private void addThingies(WorkRoom wr, JSONObject jsonObject) {
        JSONArray jsonArray = jsonObject.getJSONArray("thingies");
        for (Object json : jsonArray) {
            JSONObject nextThingy = (JSONObject) json;
            addThingy(wr, nextThingy);
        }
    }

    // MODIFIES: wr
    // EFFECTS: parses thingy from JSON object and adds it to workroom
    private void addThingy(WorkRoom wr, JSONObject jsonObject) {
        String name = jsonObject.getString("name");
        Category category = Category.valueOf(jsonObject.getString("category"));
        Thingy thingy = new Thingy(name, category);
        wr.addThingy(thingy);
    }
     */





}
